# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "TMS integration service — proxies freight load operations to the Turvo external API.",
        "title": "Drumkit Integration API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/integrations/webhooks/loads": {
            "post": {
                "description": "Receives a freight load payload and creates it in Turvo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loads"
                ],
                "summary": "Create a load via webhook",
                "parameters": [
                    {
                        "description": "Load payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLoadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLoadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/loads": {
            "get": {
                "description": "Returns all freight loads synced from Turvo.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loads"
                ],
                "summary": "List all loads",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.LoadResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CarrierDTO": {
            "type": "object",
            "properties": {
                "confirmationReceivedTime": {
                    "type": "string"
                },
                "confirmationSentTime": {
                    "type": "string"
                },
                "deliveryEnd": {
                    "type": "string"
                },
                "deliveryStart": {
                    "type": "string"
                },
                "dispatchCity": {
                    "type": "string"
                },
                "dispatchState": {
                    "type": "string"
                },
                "dispatchedTime": {
                    "type": "string"
                },
                "dispatcher": {
                    "type": "string"
                },
                "dotNumber": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expectedDeliveryTime": {
                    "type": "string"
                },
                "expectedPickupTime": {
                    "type": "string"
                },
                "externalTMSId": {
                    "type": "string"
                },
                "externalTMSTrailerId": {
                    "type": "string"
                },
                "externalTMSTruckId": {
                    "type": "string"
                },
                "firstDriverName": {
                    "type": "string"
                },
                "firstDriverPhone": {
                    "type": "string"
                },
                "mcNumber": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "pickupEnd": {
                    "type": "string"
                },
                "pickupStart": {
                    "type": "string"
                },
                "scac": {
                    "type": "string"
                },
                "sealNumber": {
                    "type": "string"
                },
                "secondDriverName": {
                    "type": "string"
                },
                "secondDriverPhone": {
                    "type": "string"
                },
                "signedBy": {
                    "type": "string"
                }
            }
        },
        "dto.CreateLoadRequest": {
            "type": "object",
            "required": [
                "consignee",
                "customer",
                "freightLoadID",
                "pickup",
                "status"
            ],
            "properties": {
                "billTo": {
                    "$ref": "#/definitions/dto.PartyDTO"
                },
                "billableWeight": {
                    "type": "number"
                },
                "carrier": {
                    "$ref": "#/definitions/dto.CarrierDTO"
                },
                "consignee": {
                    "$ref": "#/definitions/dto.StopPartyDTO"
                },
                "customer": {
                    "$ref": "#/definitions/dto.PartyDTO"
                },
                "freightLoadID": {
                    "type": "string"
                },
                "inPalletCount": {
                    "type": "integer"
                },
                "numCommodities": {
                    "type": "integer"
                },
                "operator": {
                    "type": "string"
                },
                "outPalletCount": {
                    "type": "integer"
                },
                "pickup": {
                    "$ref": "#/definitions/dto.StopPartyDTO"
                },
                "poNums": {
                    "type": "string"
                },
                "rateData": {
                    "$ref": "#/definitions/dto.RateDataDTO"
                },
                "routeMiles": {
                    "type": "number"
                },
                "specifications": {
                    "$ref": "#/definitions/dto.SpecificationsDTO"
                },
                "status": {
                    "type": "string"
                },
                "totalWeight": {
                    "type": "number"
                }
            }
        },
        "dto.CreateLoadResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "dto.LoadResponse": {
            "type": "object",
            "properties": {
                "billTo": {
                    "$ref": "#/definitions/dto.PartyDTO"
                },
                "billableWeight": {
                    "type": "number"
                },
                "carrier": {
                    "$ref": "#/definitions/dto.CarrierDTO"
                },
                "consignee": {
                    "$ref": "#/definitions/dto.StopPartyDTO"
                },
                "customer": {
                    "$ref": "#/definitions/dto.PartyDTO"
                },
                "externalTMSLoadID": {
                    "type": "string"
                },
                "freightLoadID": {
                    "type": "string"
                },
                "inPalletCount": {
                    "type": "integer"
                },
                "numCommodities": {
                    "type": "integer"
                },
                "operator": {
                    "type": "string"
                },
                "outPalletCount": {
                    "type": "integer"
                },
                "pickup": {
                    "$ref": "#/definitions/dto.StopPartyDTO"
                },
                "poNums": {
                    "type": "string"
                },
                "rateData": {
                    "$ref": "#/definitions/dto.RateDataDTO"
                },
                "routeMiles": {
                    "type": "number"
                },
                "specifications": {
                    "$ref": "#/definitions/dto.SpecificationsDTO"
                },
                "status": {
                    "type": "string"
                },
                "totalWeight": {
                    "type": "number"
                }
            }
        },
        "dto.PartyDTO": {
            "type": "object",
            "properties": {
                "addressLine1": {
                    "type": "string"
                },
                "addressLine2": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "externalTMSId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "refNumber": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "zipcode": {
                    "type": "string"
                }
            }
        },
        "dto.RateDataDTO": {
            "type": "object",
            "properties": {
                "carrierLhRateUsd": {
                    "type": "number"
                },
                "carrierMaxRate": {
                    "type": "number"
                },
                "carrierNumHours": {
                    "type": "number"
                },
                "carrierRateType": {
                    "type": "string"
                },
                "customerLhRateUsd": {
                    "type": "number"
                },
                "customerNumHours": {
                    "type": "number"
                },
                "customerRateType": {
                    "type": "string"
                },
                "fscPerMile": {
                    "type": "number"
                },
                "fscPercent": {
                    "type": "number"
                },
                "netProfitUsd": {
                    "type": "number"
                },
                "profitPercent": {
                    "type": "number"
                }
            }
        },
        "dto.SpecificationsDTO": {
            "type": "object",
            "properties": {
                "customBonded": {
                    "type": "boolean"
                },
                "escorts": {
                    "type": "boolean"
                },
                "hazmat": {
                    "type": "boolean"
                },
                "insideDelivery": {
                    "type": "boolean"
                },
                "insidePickup": {
                    "type": "boolean"
                },
                "labor": {
                    "type": "boolean"
                },
                "liftgateDelivery": {
                    "type": "boolean"
                },
                "liftgatePickup": {
                    "type": "boolean"
                },
                "maxTempFahrenheit": {
                    "type": "number"
                },
                "minTempFahrenheit": {
                    "type": "number"
                },
                "oversized": {
                    "type": "boolean"
                },
                "permits": {
                    "type": "boolean"
                },
                "seal": {
                    "type": "boolean"
                },
                "straps": {
                    "type": "boolean"
                },
                "tarps": {
                    "type": "boolean"
                }
            }
        },
        "dto.StopPartyDTO": {
            "type": "object",
            "properties": {
                "addressLine1": {
                    "type": "string"
                },
                "addressLine2": {
                    "type": "string"
                },
                "apptNote": {
                    "type": "string"
                },
                "apptTime": {
                    "type": "string"
                },
                "businessHours": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "externalTMSId": {
                    "type": "string"
                },
                "mustDeliver": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "readyTime": {
                    "type": "string"
                },
                "refNumber": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "warehouseId": {
                    "type": "string"
                },
                "zipcode": {
                    "type": "string"
                }
            }
        }
    }
}